+++ Ruddarr/Views/Movies/MovieDetails+Overview.swift
@@ -41,19 +41,16 @@ extension MovieDetails {
             .font(.caption)
             .fontWeight(.semibold)
             .textCase(.uppercase)
-            .shimmering(active: isDownloading, color: settings.theme.tint)
+            .shimmering(active: queueStatus != nil, color: settings.theme.tint)
     }
 
     var stateLabel: String {
-        if isDownloading {
-            return String(localized: "Downloading")
-        }
-
+        if let label = queueStatus?.label { return label }
         return movie.stateLabel
     }
 
-    var isDownloading: Bool {
-        queue.isDownloading(movie, instanceId: instance.id)
+    var queueStatus: QueueItemStatus? {
+        queue.queueStatus(movie, instanceId: instance.id)
     }
 
     var detailsTitle: some View {