+++ Ruddarr/Views/Series/SeriesDetails+Overview.swift
@@ -42,19 +42,16 @@ extension SeriesDetails {
             .font(.caption)
             .fontWeight(.semibold)
             .textCase(.uppercase)
-            .shimmering(active: isDownloading, color: settings.theme.tint)
+            .shimmering(active: queueStatus != nil, color: settings.theme.tint)
     }
 
-    var stateLabel: LocalizedStringKey {
-        if isDownloading {
-            return "Downloading"
-        }
-
+    var stateLabel: String {
+        if let label = queueStatus?.label { return label }
         return series.stateLabel
     }
 
-    var isDownloading: Bool {
-        queue.isDownloading(series, instanceId: instance.id)
+    var queueStatus: QueueItemStatus? {
+        queue.queueStatus(series, instanceId: instance.id)
     }
 
     var detailsTitle: some View {